+++ handlers/oauth.go
@@ -6,17 +6,14 @@ import (
 	"github.com/statping/statping/types/core"
 	"github.com/statping/statping/types/null"
 	"github.com/statping/statping/types/users"
+	"golang.org/x/oauth2"
 	"net/http"
 )
 
 type oAuth struct {
-	ID           string
-	Email        string
-	Username     string
-	Token        string
-	RefreshToken string
-	Valid        bool
-	Type         string
+	Email    string
+	Username string
+	*oauth2.Token
 }
 
 func oauthHandler(w http.ResponseWriter, r *http.Request) {
@@ -51,9 +48,7 @@ func oauthLogin(oauth *oAuth, w http.ResponseWriter, r *http.Request) {
 		Admin:    null.NewNullBool(true),
 	}
 	log.Infoln(fmt.Sprintf("OAuth %s User %s logged in from IP %s", oauth.Type, oauth.Email, r.RemoteAddr))
-
 	setJwtToken(user, w)
 
-	//returnJson(user, w, r)
 	http.Redirect(w, r, core.App.Domain+"/dashboard", http.StatusPermanentRedirect)
 }