+++ handlers/oauth_slack.go
@@ -9,18 +9,19 @@ import (
 	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/slack"
 	"net/http"
+	"strings"
 	"time"
 )
 
 func slackOAuth(r *http.Request) (*oAuth, error) {
-	c := core.App
+	auth := core.App.OAuth
 	code := r.URL.Query().Get("code")
 
 	config := &oauth2.Config{
-		ClientID:     c.OAuth.SlackClientID,
-		ClientSecret: c.OAuth.SlackClientSecret,
+		ClientID:     auth.SlackClientID,
+		ClientSecret: auth.SlackClientSecret,
 		Endpoint:     slack.Endpoint,
-		RedirectURL:  c.Domain + basePath + "oauth/slack",
+		RedirectURL:  core.App.Domain + basePath + "oauth/slack",
 		Scopes:       []string{"identity.basic"},
 	}
 
@@ -29,11 +30,8 @@ func slackOAuth(r *http.Request) (*oAuth, error) {
 		return nil, err
 	}
 
-	oauther := &oAuth{
-		Token:        gg.AccessToken,
-		RefreshToken: gg.RefreshToken,
-		Valid:        gg.Valid(),
-		Type:         gg.Type(),
+	if !gg.Valid() {
+		return nil, errors.New("oauth token is not valid")
 	}
 
 	identity, err := returnSlackIdentity(gg.AccessToken)
@@ -45,10 +43,36 @@ func slackOAuth(r *http.Request) (*oAuth, error) {
 		return nil, errors.New("slack identity is invalid")
 	}
 
-	oauther.Username = identity.User.Name
-	oauther.Email = identity.User.Email
+	if !validateSlack(identity) {
+		return nil, errors.New("slack user is not whitelisted")
+	}
 
-	return oauther, nil
+	return &oAuth{
+		Token:    gg,
+		Username: strings.ToLower(identity.User.Name),
+		Email:    strings.ToLower(identity.User.Email),
+	}, nil
+}
+
+func validateSlack(id slackIdentity) bool {
+	auth := core.App.OAuth
+	if auth.SlackUsers == "" {
+		return true
+	}
+
+	if auth.SlackUsers != "" {
+		users := strings.Split(auth.SlackUsers, ",")
+		for _, u := range users {
+			if strings.ToLower(u) == strings.ToLower(id.User.Email) {
+				return true
+			}
+			if strings.ToLower(u) == strings.ToLower(id.User.Name) {
+				return true
+			}
+		}
+	}
+
+	return false
 }
 
 // slackIdentity will query the Slack API to fetch the users ID, username, and email address.