+++ notifiers/webhook.go
@@ -127,7 +127,7 @@ func (w *webhooker) OnTest() (string, error) {
 }
 
 // OnFailure will trigger failing service
-func (w *webhooker) OnFailure(s *services.Service, f *failures.Failure) (string, error) {
+func (w *webhooker) OnFailure(s services.Service, f failures.Failure) (string, error) {
 	msg := ReplaceVars(w.FailureData, s, f)
 	resp, err := w.sendHttpWebhook(msg)
 	if err != nil {
@@ -139,8 +139,8 @@ func (w *webhooker) OnFailure(s *services.Service, f *failures.Failure) (string,
 }
 
 // OnSuccess will trigger successful service
-func (w *webhooker) OnSuccess(s *services.Service) (string, error) {
-	msg := ReplaceVars(w.SuccessData, s, nil)
+func (w *webhooker) OnSuccess(s services.Service) (string, error) {
+	msg := ReplaceVars(w.SuccessData, s, failures.Failure{})
 	resp, err := w.sendHttpWebhook(msg)
 	if err != nil {
 		return "", err