+++ types/services/failures.go
@@ -20,15 +20,14 @@ func (s *Service) FailuresSince(t time.Time) failures.Failurer {
 	return failures.Since(t, s)
 }
 
-func (s *Service) DowntimeAgo() string {
-	last := s.LastOnline
-	if last.IsZero() {
+func (s Service) DowntimeAgo() string {
+	if s.LastOnline.IsZero() {
 		return "Never been online"
 	}
-	return humanize.Time(last)
+	return humanize.Time(s.LastOnline)
 }
 
-func (s *Service) DowntimeText() string {
+func (s Service) DowntimeText() string {
 	last := s.AllFailures().Last()
 	if last == nil {
 		return ""